requirements.txt
gunicorn==20.1.0
Flask==2.3.3
requests==2.31.0
google-cloud-storage==2.14.0
pdf2image==1.16.3
sympy==1.13.3
numpy==2.2.4
regex==2024.11.6
google-generativeai==0.3.2 